noc-product/pyproject.toml
[project]
name = "noc-product"
version = "0.1.0"
description = "4WRD NOC Incident Management — E6a product build"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "itsdangerous>=2.2",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "httpx>=0.27",
    "sqlalchemy>=2.0",
    "psycopg[binary,pool]>=3.1",
    "alembic>=1.13",
    "asyncpg>=0.29",
    "pgvector>=0.2.5",
    "langgraph>=0.0.40",
    "pyyaml>=6.0",
    "python-jose[cryptography]>=3.3",
    "authlib>=1.3",
    "minio>=7.2",
    "hvac>=2.1",
    "prometheus-client>=0.20",
    "opentelemetry-api>=1.24",
    "opentelemetry-sdk>=1.24",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "tenacity>=8.2",
    "structlog>=24.1",
    "unstructured[pdf]>=0.13",
    "pdfminer.six>=20231228",
    "pytesseract>=0.3.10",
    "python-docx>=1.1",
    "tiktoken>=0.6",
    "click>=8.1",
    "filelock>=3.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "hypothesis>=6.100",
    "testcontainers>=4.3",
    "ruff>=0.4",
    "mypy>=1.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["services"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
